Cloud Function/ipl_unzip_data/requirements.txt
# Function dependencies, for example:
# package>=version
google.cloud.storage